Documentation for the Test Cases
The following contains the file with the implementation of the test cases for the posts page. Additionally, a screenshot showcases the results when I run the Python file. 
Note: One test failed as I'm still working on a feature for the users to add a 'Skills Overview' section to their posts with the rest of the functionality working as intended.